requirements.txt
awscli~=1.18.69
pytz
pyOpenSSL
WTForms==2.3.3
ndg-httpsclient
jinja2~=2.8
boto3==1.16.63
qds-sdk~=1.10
pathlib~=1.0
psycopg2~=2.7
google-api-python-client~=1.7
oauth2client~=4.1
configparser~=3.5
future~=0.16
text_unidecode~=1.2
numpy==1.19.5
selenium~=3.141
funcy~=1.11
pandas~=0.24.2
pyarrow~=0.12.1
psutil~=4.4
geoip2~=2.9
geojson~=2.4
shapely~=1.6
delphi_epidata~=0.0.4
PyPDF2~=1.26
hubspot3~=3.2.50
marshmallow-sqlalchemy~=0.22.0
redis==3.2
epiweeks==2.1.2
SQLAlchemy==1.3.23
simple-salesforce==1.11.1
facebook_business
pytest
pysftp
apache-airflow[mysql,postgres,qds,s3,slack]==2.2.4
jupyter==1.0.0
xlrd~=1.1